Rename Lindent to Lindent-longlines, do not wrap lines
Line wrapping code in GNU indent is not perfect. For instance,
`indent` may wrap long lines at `.` and `->` structure member operators,
or between the word `struct` and the name of the structure.

Lindent-longlines script goal is to de-obfuscate legacy code. Line
wrapping may be done manually or not at all.